requirements.txt
streamlit
python-pptx
openai
python-dotenv
